feat: least surprise in failures
This change covers two possible failures:

* a developer requesting a chain that hasn't been configured
* network failures on subgraphs

If a developer forgets to add to the client map, this will throw an
error instead of defaulting to ethereum results. This helps the
developer make changes correctly.

If a network failure occurs or a developer makes a mistake, the entire
membership structure is not affected via `Promise.allSettled`. This avoids
confusing errors for users.